Initial commit.
Work has been going on for a little while without source control, so the
current code appears magically out of nowhere. The whole thing is fairly
close to a working initial version: we generate pages for spells, and a
spell-list for each class, ordered by spell level and spell name. We still
need:

* Css to make the pages less ugly.
* A title page
* A table of contents
* To zip everything together into a nice epub.